Move preview controllers into DesignTime folder
- Move preview controllers into new folder `DesignTime`
- Remove duplicate implementations of preview controllers
