+++ ViewKit/Sources/DesignTime/ApplicationPreviewProvider.swift
@@ -1,0 +1,5 @@
+import ModelKit
+
+final class ApplicationPreviewProvider: StateController {
+  let state = [Application]()
+}
