+++ ViewKit/Sources/DesignTime/CommandPreviewController.swift
@@ -1,0 +1,6 @@
+import ModelKit
+
+final class CommandPreviewController: ViewController {
+  let state = ModelFactory().workflowDetail().commands
+  func perform(_ action: CommandListView.Action) {}
+}
