+++ ViewKit/Sources/DesignTime/GroupPreviewController.swift
@@ -1,0 +1,6 @@
+import ModelKit
+
+final class GroupPreviewController: ViewController {
+  let state = ModelFactory().groupList()
+  func perform(_ action: GroupList.Action) {}
+}
