+++ ViewKit/Sources/DesignTime/KeyboardShortcutPreviewController.swift
@@ -1,0 +1,6 @@
+import ModelKit
+
+final class KeyboardShortcutPreviewController: ViewController {
+  let state: [ModelKit.KeyboardShortcut] = ModelFactory().keyboardShortcuts()
+  func perform(_ action: KeyboardShortcutListView.Action) {}
+}
