+++ ViewKit/Sources/DesignTime/OpenPanelPreviewController.swift
@@ -1,0 +1,6 @@
+import ModelKit
+
+final class OpenPanelPreviewController: ViewController {
+  let state = ""
+  func perform(_ action: OpenPanelAction) {}
+}
