+++ ViewKit/Sources/DesignTime/WorkflowPreviewController.swift
@@ -1,0 +1,7 @@
+import ModelKit
+
+final class WorkflowPreviewController: ViewController {
+  let state = ModelFactory().workflowList().first
+  func perform(_ action: WorkflowList.Action) {}
+}
+
