+++ ViewKit/Sources/Views/EditCommand/EditAppleScriptView.swift
@@ -55,8 +55,3 @@ struct EditAppleScriptView_Previews: PreviewProvider, TestPreviewProvider {
         openPanelController: OpenPanelPreviewController().erase())
     }
 }
-
-private final class OpenPanelPreviewController: ViewController {
-  let state = ""
-  func perform(_ action: OpenPanelAction) {}
-}
