+++ ViewKit/Sources/Views/EditCommand/EditShellScriptView.swift
@@ -55,8 +55,3 @@ struct EditShellScriptView_Previews: PreviewProvider, TestPreviewProvider {
       openPanelController: OpenPanelPreviewController().erase())
   }
 }
-
-private final class OpenPanelPreviewController: ViewController {
-  let state = ""
-  func perform(_ action: OpenPanelAction) {}
-}
