+++ ViewKit/Sources/Views/KeyboardShortcuts/KeyboardShortcutListView.swift
@@ -65,8 +65,3 @@ struct KeyboardShortcutListView_Previews: PreviewProvider, TestPreviewProvider {
     KeyboardShortcutListView(keyboardShortcutController: KeyboardShortcutPreviewController().erase())
   }
 }
-
-private final class KeyboardShortcutPreviewController: ViewController {
-  let state: [ModelKit.KeyboardShortcut] = ModelFactory().keyboardShortcuts()
-  func perform(_ action: KeyboardShortcutListView.Action) {}
-}
