+++ ViewKit/Sources/Views/Workflow/WorkflowView.swift
@@ -133,22 +133,3 @@ struct WorkflowView_Previews: PreviewProvider, TestPreviewProvider {
       .frame(height: 668)
   }
 }
-
-private final class ApplicationPreviewProvider: StateController {
-  let state = [Application]()
-}
-
-private final class CommandPreviewController: ViewController {
-  let state = ModelFactory().workflowDetail().commands
-  func perform(_ action: CommandListView.Action) {}
-}
-
-private final class KeyboardShortcutPreviewController: ViewController {
-  let state: [ModelKit.KeyboardShortcut] = ModelFactory().keyboardShortcuts()
-  func perform(_ action: KeyboardShortcutListView.Action) {}
-}
-
-private final class OpenPanelPreviewController: ViewController {
-  let state = ""
-  func perform(_ action: OpenPanelAction) {}
-}
