+++ ViewKit/Sources/Views/WorkflowList/WorkflowList.swift
@@ -84,8 +84,3 @@ struct WorkflowList_Previews: PreviewProvider, TestPreviewProvider {
       .environmentObject(UserSelection())
   }
 }
-
-private final class WorkflowPreviewController: ViewController {
-  let state = ModelFactory().workflowList().first
-  func perform(_ action: WorkflowList.Action) {}
-}
